@@ -21,57 +22,40 @@ apply plugin: 'com.neenbedankt.android-apt'
  */
 
 // =============================================================================
-// https://developer.android.com/studio/publish/app-signing.html#secure-shared-keystore
-
-// Create a variable called keystorePropertiesFile, and initialize it to your
-// keystore.properties file, in the rootProject folder.
-def keystorePropertiesFile = rootProject.file("keystore.properties")
-
-// Initialize a new Properties() object called keystoreProperties.
-def keystoreProperties = new Properties()
-
-// Load your keystore.properties file into the keystoreProperties object.
-keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
-
-// =============================================================================
 
 android {
+    compileSdkVersion 32
+
     signingConfigs {
         config {
-            keyAlias keystoreProperties['keyAlias']
-            keyPassword keystoreProperties['keyPassword']
-            storeFile file(keystoreProperties['storeFile'])
-            storePassword keystoreProperties['storePassword']
+            storeFile file('chatcert2.jks')
+            keyAlias 'cheng'
+            keyPassword 'android'
+            storePassword 'android'
+            v1SigningEnabled true
+            v2SigningEnabled true
         }
     }
-    compileSdkVersion 25
-    buildToolsVersion "24.0.3"
+
     defaultConfig {
         applicationId 'com.philliphsu.clock2'
-        minSdkVersion 19
-        targetSdkVersion 25
-        versionCode 113
-        versionName "1.1.3"
-        // Disabled for now because we're not ready to
-        // completely port over to vector drawables
-//        vectorDrawables.useSupportLibrary = true
+        minSdkVersion MIN_SDK_VERSION as int
+        targetSdkVersion TARGET_SDK_VERSION as int
+
+        versionCode VERSION_CODE as int
+        versionName VERSION_NAME
     }
     buildTypes {
-        release {
-            // https://developer.android.com/studio/build/shrink-code.html#shrink-code
-            //
-            // Proguard is disabled, because it seems like it is removing
-            // ButterKnife generated code and I don't know how to fix it...
-            minifyEnabled false
-            // "'proguard-android-optimize.txt' includes the same ProGuard rules
-            // [as 'proguard-android.txt'], but with other optimizations that
-            // perform analysis at the bytecode level—inside and across methods—
-            // to reduce your APK size further and help it run faster."
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-            signingConfig signingConfigs.config
-        }
         debug {
-            applicationIdSuffix ".debug"
+            buildConfigField "boolean", "LOG_DEBUG", "true"
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+
+        release {
+            buildConfigField "boolean", "LOG_DEBUG", "false"
+            minifyEnabled true
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
     productFlavors {
