@@ -20,8 +20,9 @@
 package com.philliphsu.clock2.dialogs;
 
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.AppCompatEditText;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.widget.AppCompatEditText;
+
 import android.view.inputmethod.EditorInfo;
 import android.widget.EditText;
 
