@@ -24,8 +24,9 @@ import android.database.Cursor;
 import android.media.RingtoneManager;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AlertDialog;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+
 
 import com.philliphsu.clock2.R;
 import com.philliphsu.clock2.ringtone.playback.RingtoneLoop;
