@@ -20,7 +20,7 @@
 package com.philliphsu.clock2.dialogs;
 
 import android.net.Uri;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentManager;
 import android.util.Log;
 
 import com.philliphsu.clock2.dialogs.DialogFragmentController;
