@@ -25,7 +25,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Parcelable;
 import android.os.SystemClock;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 import android.view.ViewGroup;
 
 import com.philliphsu.clock2.ringtone.playback.RingtoneService;
