@@ -25,7 +25,7 @@ import android.net.Uri;
 import android.os.Parcelable;
 import android.preference.PreferenceManager;
 import android.provider.Settings;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 
 import com.philliphsu.clock2.R;
 import com.philliphsu.clock2.timers.Timer;
