@@ -27,7 +27,7 @@ import android.net.Uri;
 import android.os.Parcelable;
 import android.preference.RingtonePreference;
 import android.provider.Settings;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.AttributeSet;
 
 import com.philliphsu.clock2.dialogs.RingtonePickerDialog;
