@@ -20,8 +20,8 @@
 package com.philliphsu.clock2.util;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
 
 import com.philliphsu.clock2.R;
 
