@@ -20,9 +20,10 @@
 package com.philliphsu.clock2;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatSeekBar;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
+
+import androidx.appcompat.widget.AppCompatSeekBar;
 
 /**
  * A SeekBar that cannot be touch controlled.
