@@ -28,7 +28,7 @@ import com.philliphsu.clock2.alarms.Alarm;
 import com.philliphsu.clock2.alarms.misc.AlarmController;
 import com.philliphsu.clock2.alarms.misc.DaysOfWeek;
 
-import butterknife.Bind;
+import butterknife.BindView;
 
 import static android.view.View.GONE;
 import static android.view.View.VISIBLE;
@@ -39,8 +39,8 @@ import static com.philliphsu.clock2.alarms.misc.DaysOfWeek.NUM_DAYS;
  */
 public class CollapsedAlarmViewHolder extends BaseAlarmViewHolder {
 
-    @Bind(R.id.countdown) AlarmCountdown mCountdown;
-    @Bind(R.id.recurring_days) TextView mDays; // TODO: use `new DateFormatSymbols().getShortWeekdays()` to set texts
+    @BindView(R.id.countdown) AlarmCountdown mCountdown;
+    @BindView(R.id.recurring_days) TextView mDays; // TODO: use `new DateFormatSymbols().getShortWeekdays()` to set texts
 
     public CollapsedAlarmViewHolder(ViewGroup parent, OnListItemInteractionListener<Alarm> listener,
                                     AlarmController alarmController) {
