@@ -19,7 +19,7 @@
 
 package com.philliphsu.clock2.util;
 
-import android.support.design.widget.Snackbar;
+import com.google.android.material.snackbar.Snackbar;
 import android.view.View;
 
 /**
